requirements.txt
certifi==2021.5.30
chardet==4.0.0
boto3==1.18.53
botocore==1.21.53
dj-database-url==0.5.0
django-cloudinary-storage==0.3.0
cloudinary==1.26.0
folium==0.12.1.post1
Django==3.2.4
django-crispy-forms==1.12.0
django-heroku==0.3.1
django-storages==1.11.1
docutils==0.17.1
gunicorn==20.1.0
idna==2.10
jmespath==0.10.0
Pillow==8.3.2
psycopg2==2.9.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.5.0
six==1.16.0
urllib3==1.26.5
whitenoise==5.3.0

